Move utxo db operations to its own file
