+++ blockchain_storage/src/utxo_db.rs
@@ -1,14 +1,10 @@
 #![allow(dead_code)]
 
-use std::collections::HashMap;
-
 use crate::{BlockchainStorageRead, BlockchainStorageWrite, Error, Store, UtxoRead, UtxoWrite};
 use common::chain::block::Block;
 use common::chain::OutPoint;
-use common::primitives::{Id, H256};
-use utxo::{
-    FlushableUtxoView, OutPointKey, Utxo, UtxoEntry, UtxosCache, UtxosPersistentStorage, UtxosView,
-};
+use common::primitives::Id;
+use utxo::{utxo_storage::UtxosPersistentStorage, Utxo};
 
 pub struct UtxoDBInterface {
     store: Store,
@@ -40,78 +36,9 @@ impl UtxosPersistentStorage for UtxoDBInterface {
     }
 }
 
-pub struct UtxoDB<'a, S: UtxosPersistentStorage>(&'a mut S);
-
 impl From<Error> for utxo::Error {
     fn from(e: Error) -> Self {
         utxo::Error::DBError(format!("{:?}", e))
-    }
-}
-
-impl<'a, S: UtxosPersistentStorage> UtxoDB<'a, S> {
-    pub fn new(store: &'a mut S) -> Self {
-        Self(store)
-    }
-}
-
-impl<'a, S: UtxosPersistentStorage> UtxosView for UtxoDB<'a, S> {
-    fn get_utxo(&self, outpoint: &OutPoint) -> Option<Utxo> {
-        match self.0.get_utxo(outpoint) {
-            Ok(res) => res,
-            Err(_) => {
-                //todo: handle errors
-                None
-            }
-        }
-    }
-
-    fn has_utxo(&self, outpoint: &OutPoint) -> bool {
-        self.get_utxo(outpoint).is_some()
-    }
-
-    fn get_best_block_hash(&self) -> Option<H256> {
-        match self.0.get_best_block_id() {
-            Ok(opt_id) => opt_id.map(|id| id.get()),
-            Err(_) => {
-                // TODO: handle errors
-                None
-            }
-        }
-    }
-
-    fn derive_cache(&self) -> UtxosCache {
-        let mut cache = UtxosCache::new(self);
-        if let Some(hash) = self.get_best_block_hash() {
-            cache.set_best_block(hash);
-        }
-        cache
-    }
-
-    fn estimated_size(&self) -> usize {
-        todo!()
-    }
-}
-
-impl<'a, S: UtxosPersistentStorage> FlushableUtxoView for UtxoDB<'a, S> {
-    fn batch_write(
-        &mut self,
-        utxos: HashMap<OutPointKey, UtxoEntry>,
-        block_hash: H256,
-    ) -> Result<(), utxo::Error> {
-        // check each entry if it's dirty. Only then will the db be updated.
-        for (key, entry) in utxos {
-            let outpoint: OutPoint = (&key).into();
-            if entry.is_dirty() {
-                if let Some(utxo) = entry.utxo() {
-                    self.0.set_utxo(&outpoint, utxo)?;
-                } else {
-                    // entry is spent
-                    self.0.del_utxo(&outpoint)?;
-                }
-            }
-        }
-        self.0.set_best_block_id(&Id::<Block>::new(&block_hash))?;
-        Ok(())
     }
 }
 
@@ -120,9 +47,15 @@ mod test {
     use super::*;
     use common::chain::{Destination, OutPointSourceId, Transaction, TxOutput};
     use common::primitives::{Amount, BlockHeight};
+    use common::primitives::{Id, H256};
     use iter_tools::Itertools;
     use rand::Rng;
+    use std::collections::HashMap;
     use utxo::flush_to_base;
+    use utxo::{
+        utxo_storage::UtxoDB, FlushableUtxoView, OutPointKey, Utxo, UtxoEntry, UtxosCache,
+        UtxosView,
+    };
 
     fn create_utxo(block_height: u64) -> (Utxo, OutPoint) {
         let output = TxOutput::new(Amount::new(10), Destination::PublicKey);
